LogTransaction Fix: remove "Default" from file name
