--- src/java/org/apache/cassandra/db/lifecycle/DefaultLogTransactionsFactory.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.cassandra.db.lifecycle;
-
-import org.apache.cassandra.db.compaction.OperationType;
-import org.apache.cassandra.schema.TableMetadataRef;
-
-final class DefaultLogTransactionsFactory implements LogTransactionsFactory
-{
-    @Override
-    public LogTransaction createLogTransaction(OperationType operationType, TableMetadataRef metadata)
-    {
-        logger.debug("Creating a transaction for {} on {}", operationType, metadata);
-        return new DefaultLogTransaction(operationType);
-    }
-
-    @Override
-    public LogAwareFileLister createLogAwareFileLister()
-    {
-        return new DefaultLogAwareFileLister();
-    }
-
-    @Override
-    public LogFileCleaner createLogFileCleaner()
-    {
-        return new DefaultLogFileCleaner();
-    }
-
-    @Override
-    public FailedTransactionDeletionHandler createFailedTransactionDeletionHandler()
-    {
-        return DefaultLogTransaction::rescheduleFailedDeletions;
-    }
-}
