+++ src/java/org/apache/cassandra/db/lifecycle/LogReplica.java
@@ -42,7 +42,7 @@ import org.apache.cassandra.utils.INativeLibrary;
  * partial records in case we crashed after writing to one replica but
  * before compliting the write to another replica.
  *
- * Note: this is used by {@link DefaultLogTransaction}
+ * Note: this is used by {@link LogTransaction}
  *
  * @see LogFile
  */
@@ -133,7 +133,7 @@ final class LogReplica implements AutoCloseable
 
     void delete()
     {
-        DefaultLogTransaction.delete(file);
+        LogTransaction.delete(file);
         syncDirectory();
     }
