+++ src/java/org/apache/cassandra/db/lifecycle/LogReplicaSet.java
@@ -41,7 +41,7 @@ import org.apache.cassandra.utils.Throwables;
  * A set of log replicas. This class mostly iterates over replicas when writing or reading,
  * ensuring consistency among them and hiding replication details from LogFile.
  *
- * Note: this is used by {@link DefaultLogTransaction}
+ * Note: this is used by {@link LogTransaction}
  *
  * @see LogReplica
  * @see LogFile
